requirements.txt
Django==5.2.8
reportlab==4.0.9
python-decouple==3.8
djangorestframework==3.15.2
django-filter==24.3

